Changes added

+++ backend/index.js
@@ -8,6 +8,8 @@ const app = express();
 const PORT = 5000;
 const mongoose = require("mongoose");
 const blogModel = require("./models/blogModel");
+var cookieParser = require("cookie-parser");
+const jwt = require("jsonwebtoken");
 const MONGODB_URI =
   "mongodb+srv://dragkamal71:nJpsGN4A1pHn9OFF@cluster0.zub33.mongodb.net/";
 
@@ -25,13 +27,16 @@ const connection = async () => {
 };
 connection();
 const corsOptions = {
-  origin: "*", // Allow frontend to access the backend
+  origin: "http://localhost:5173", // Allow frontend to access the backend
   methods: "GET,POST,PUT,DELETE", // Specify allowed HTTP methods
   allowedHeaders: "Content-Type,Authorization", // Specify allowed headers
+  credentials: true,
 };
 
 // Use CORS middleware
 app.use(cors(corsOptions));
+app.use(express.json());
+app.use(cookieParser());
 // Configure Multer with disk storage
 const upload = multer({
   storage: multer.diskStorage({
@@ -48,10 +53,17 @@ app.listen(PORT, () => {
   console.log(`Server running at http://localhost:${PORT}`);
 });
 app.post("/api/createData", upload.single("image"), async (req, res) => {
-  const { title, content } = req.body;
+
+  const token = jwt.verify(req.cookies.admin_token, "secret");
+  console.log(token);
+  if (!token) {
+    return res.status(401).json({ message: "Invalid token" });
+  }
+
+  const { title, content , blogType } = req.body;
 
   if (!req.file) {
-    await blogModel.create({ title, content });
+    await blogModel.create({ title, content , blogType });
 
     return res.status(200).send({ message: "No file uploaded." });
   }
@@ -67,6 +79,7 @@ app.post("/api/createData", upload.single("image"), async (req, res) => {
     await blogModel.create({
       title,
       content,
+      blogType,
       image: uploadResponse.secure_url,
     });
 
@@ -95,7 +108,7 @@ app.post("/api/createData", upload.single("image"), async (req, res) => {
 });
 app.get("/api/blogs", async (req, res) => {
   try {
-    const blogs = await blogModel.find();
+    const blogs = await blogModel.find({blogType: req.query.blogType});
     res.status(200).json(blogs);
   } catch (error) {
     console.error("Error in fetching blogs:", error);
@@ -114,13 +127,19 @@ app.get("/api/blogs/:id", async (req, res) => {
 });
 
 app.put("/api/blogs/:id", upload.single("image"), async (req, res) => {
+  const token = jwt.verify(req.cookies.admin_token, "secret");
+  console.log(token);
+  if (!token) {
+    return res.status(401).json({ message: "Invalid token" });
+  }
   if (!req.file) {
     try {
-      const { title, content, existimage } = req.body;
+      const { title, content, existimage, blogType } = req.body;
 
       const blog = await blogModel.findByIdAndUpdate(req.params.id, {
         title,
         content,
+        blogType,
         image: existimage,
       });
       res.status(200).json(blog);
@@ -130,7 +149,7 @@ app.put("/api/blogs/:id", upload.single("image"), async (req, res) => {
     }
   } else {
     try {
-      const { title, content } = req.body;
+      const { title, content , blogType } = req.body;
 
       const filePath = req.file.path; // Path to the uploaded file
 
@@ -142,6 +161,7 @@ app.put("/api/blogs/:id", upload.single("image"), async (req, res) => {
       const blog = await blogModel.findByIdAndUpdate(req.params.id, {
         title,
         content,
+          blogType,
         image: uploadResponse.secure_url,
       });
 
@@ -162,18 +182,48 @@ app.put("/api/blogs/:id", upload.single("image"), async (req, res) => {
 });
 
 app.delete("/api/blogs/:id", async (req, res) => {
+  console.log(req.cookies.admin_token);
   try {
+    if (!jwt.verify(req.cookies.admin_token, "secret")) {
+      return res.status(401).json({ message: "Invalid token" });
+    }
+
     const blog = await blogModel.findByIdAndDelete(req.params.id);
     res.status(200).json(blog);
-  }
-  catch(error){
+  } catch (error) {
     console.error("Error in deleting blog:", error);
     res.status(500).json({ message: "Server error while deleting blog." });
   }
-})
+});
+
+app.post("/api/login", async (req, res) => {
+  const { username, password } = req.body;
+  console.log(username, password);
+
+  if (username === "admin" && password === "Suvam@4256") {
+    const token = jwt.sign({ token: "suvamhere@4256" }, "secret");
+    res.cookie("admin_token", token);
+    return res.json({ message: "Cookie has been set" });
+  }
+  res.status(401).json({ message: "Invalid username or password" });
+});
+
+app.post("/api/logout", async (req, res) => {
+  const token = jwt.verify(req.cookies.admin_token, "secret");
+  console.log(token);
+  if (
+    req.cookies.admin_token &&
+    token
+  ) {
+    return res
+      .clearCookie("admin_token")
+      .json({ message: "Logged out successfully" });
+  }
+  res.status(401).json({ message: "Invalid token" });
+}); 
 
 // Ensure the 'uploads' directory exists
-const UPLOADS_DIR = path.join(__dirname, "./uploads");
+const UPLOADS_DIR = path.join(__dirname, "./uploads"); 
 if (!fs.existsSync(UPLOADS_DIR)) {
   fs.mkdirSync(UPLOADS_DIR);
 }
